cli: Move CLI logger out of internal
This will allow commands to be moved around without getting blocked by
the logger being inside the internal module, which is a prerequisite
for #1992.
