+++ cmd/gitops/logger/logger.go
@@ -1,4 +1,4 @@
-package internal
+package logger
 
 import (
 	"fmt"
