+++ cmd/gitops/logger/logger_test.go
@@ -1,4 +1,4 @@
-package internal
+package logger
 
 import (
 	"bytes"
