+++ cmd/gitops/version/cmd.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/weaveworks/go-checkpoint"
-	"github.com/weaveworks/weave-gitops/cmd/internal"
+	"github.com/weaveworks/weave-gitops/cmd/gitops/logger"
 
 	"github.com/spf13/cobra"
 )
@@ -34,7 +34,7 @@ func runCmd(cmd *cobra.Command, args []string) {
 
 // CheckVersion looks to see if there is a newer version of the software available
 func CheckVersion(p *checkpoint.CheckParams) {
-	log := internal.NewCLILogger(os.Stdout)
+	log := logger.NewCLILogger(os.Stdout)
 	checkResponse, err := checkpoint.Check(p)
 
 	if err == nil && checkResponse.Outdated {
